% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate_probe_data.R
\name{calibrate_probe_data}
\alias{calibrate_probe_data}
\title{calibrate probe meta data}
\usage{
calibrate_probe_data(path_row_data, probe)
}
\arguments{
\item{path_row_data}{location of soil probe data table}

\item{probe}{name of probe/ file}
}
\description{
function to calibrate the cleaned probe data
}
\examples{
\dontrun{
calibrate_probe_data
}
}
\author{
Maiken Baumberger
}
