% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_probe_data.R
\name{clean_probe_data}
\alias{clean_probe_data}
\title{clean probe meta data}
\usage{
clean_probe_data(
  path_row_data,
  probe,
  startdatetime,
  enddatetime,
  organic,
  excess
)
}
\arguments{
\item{path_row_data}{loaction of soil probe data table}

\item{probe}{name of probe/ file}

\item{startdatetime}{start time}

\item{enddatetime}{end time}

\item{organic}{depth of organic layer in cm}

\item{excess}{excess of probe in cm}
}
\description{
function to clean the probe data
}
\examples{
\dontrun{
clean_probe_data
}
}
\author{
Maiken Baumberger
}
