% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_carbon4d_data.R
\name{get_carbon4d_data}
\alias{get_carbon4d_data}
\title{get carbon4d data from github to a local folder}
\usage{
get_carbon4d_data(path)
}
\arguments{
\item{path}{path to the local saved}
}
\description{
get carbon4d data from github to a local folder
}
\examples{
\dontrun{
get_carbon4d_data
}
}
\author{
Maiken Baumberger
}
