% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_longterm_probe_data.R
\name{load_longterm_probe_data}
\alias{load_longterm_probe_data}
\title{load longterm probe data from a local folder into r}
\usage{
load_longterm_probe_data(path)
}
\arguments{
\item{path}{path to the local saved}
}
\value{

}
\description{
load longterm probe data from a local folder into r
}
\examples{
\dontrun{
load_longterm_probe_data
}
}
\author{
Maiken Baumberger
}
