% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_monthly_probe_data.R
\name{load_monthly_probe_data}
\alias{load_monthly_probe_data}
\title{load monthly probe data from a local folder into r}
\usage{
load_monthly_probe_data(path)
}
\arguments{
\item{path}{path to the local saved}
}
\value{

}
\description{
load monthly probe data from a local folder into r
}
\examples{
\dontrun{
load_monthly_probe_data
}
}
\author{
Maiken Baumberger
}
