% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_probe_meta_data.R
\name{load_probe_meta_data}
\alias{load_probe_meta_data}
\title{load probe meta data from a local folder into r}
\usage{
load_probe_meta_data(path)
}
\arguments{
\item{path}{path to the local saved}
}
\value{

}
\description{
load probe meta data from a local folder into r
}
\examples{
\dontrun{
load_probe_meta_data
}
}
\author{
Maiken Baumberger
}
