% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_probe_meta_data_longterm.R,
%   R/taunus_load_probe_meta_data.R
\name{load_probe_meta_data_longterm}
\alias{load_probe_meta_data_longterm}
\alias{taunus_load_probe_meta_data}
\title{load probe meta data longterm from a local folder into r}
\usage{
load_probe_meta_data_longterm(path)

load_probe_meta_data_longterm(path)
}
\arguments{
\item{path}{path to the local saved}
}
\description{
load probe meta data longterm from a local folder into r

load probe meta data taunus from a local folder into r
}
\examples{
\dontrun{
load_probe_meta_data_longterm
}
\dontrun{
taunus_load_probe_meta_data
}
}
\author{
Maiken Baumberger
}
