% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_probe_meta_data_monthly.R
\name{load_probe_meta_data_monthly}
\alias{load_probe_meta_data_monthly}
\title{load probe meta data monthly from a local folder into r}
\usage{
load_probe_meta_data_monthly(path)
}
\arguments{
\item{path}{path to the local saved}
}
\description{
load probe meta data monthly from a local folder into r
}
\examples{
\dontrun{
load_probe_meta_data_monthly
}
}
\author{
Maiken Baumberger
}
