% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_probe_meta_data_weekly.R
\name{load_probe_meta_data_weekly}
\alias{load_probe_meta_data_weekly}
\title{load probe meta data weekly from a local folder into r}
\usage{
load_probe_meta_data_weekly(path)
}
\arguments{
\item{path}{path to the local saved}
}
\description{
load probe meta data weekly from a local folder into r
}
\examples{
\dontrun{
load_probe_meta_data_weekly
}
}
\author{
Maiken Baumberger
}
