% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_longterm_porbe_plots.R
\name{plot_longterm_porbe_plots}
\alias{plot_longterm_porbe_plots}
\title{mapview plot of the locations of the monthly porbe data}
\usage{
plot_longterm_porbe_plots(path)
}
\arguments{
\item{path}{path to the local saved}
}
\value{
A mapview output
}
\description{
mapview plot of the locations of the longterm porbe data
}
\examples{
plot_longterm_porbe_plots(path="C:/Users/maike/Desktop/Carbon4D/DownloadGitData/ProbeMetaDataLongterm/ProbeMetaDataLongterm.csv")
\dontrun{
plot_longterm_porbe_plots()
}
}
\author{
Maiken Baumberger
}
