% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_plots.R
\name{plot_monthly_porbe_plots}
\alias{plot_monthly_porbe_plots}
\alias{plot_plots}
\title{mapview plot of the locations of the monthly porbe data}
\usage{
plot_monthly_porbe_plots(path)
}
\arguments{
\item{path}{path to the local saved}
}
\value{
A mapview output
}
\description{
mapview plot of the locations of the monthly porbe data
}
\examples{
\dontrun{
plot_plots()
}
}
\author{
Maiken Baumberger
}
