% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPlots.R
\name{plotPlots}
\alias{plotPlots}
\title{Plot the Carbon4D plots}
\usage{
plotPlots()
}
\value{
A mapview output
}
\description{
Plot the Carbon4D plots
}
\note{

}
\examples{
\dontrun{
plotPlots()
}
}
\references{

}
\author{
Hanna Meyer
}
