% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_probe_data.R
\docType{data}
\name{probe}
\alias{probe}
\alias{clean_probe_data}
\title{clean probe meta data}
\format{
An object of class \code{character} of length 1.
}
\usage{
probe
}
\arguments{
\item{path}{path to the local saved}
}
\description{
clean probe meta data
}
\examples{
\dontrun{
clean_probe_data
}
}
\author{
Maiken Baumberger
}
\keyword{datasets}
