% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taunus_load_probe_data.R
\name{taunus_load_probe_data}
\alias{taunus_load_probe_data}
\title{load taunus probe data from a local folder into r}
\usage{
taunus_load_probe_data(path)
}
\arguments{
\item{path}{path to the local saved}
}
\description{
load taunus probe data from a local folder into r
}
\examples{
\dontrun{
taunus_load_probe_data
}
}
\author{
Maiken Baumberger
}
