% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taunus_load_probe_meta_data.R
\name{taunus_load_probe_meta_data}
\alias{taunus_load_probe_meta_data}
\title{load probe meta data taunus from a local folder into r}
\usage{
taunus_load_probe_meta_data(path)
}
\arguments{
\item{path}{path to the local saved}
}
\description{
load probe meta data taunus from a local folder into r
}
\examples{
\dontrun{
taunus_load_probe_meta_data
}
}
\author{
Maiken Baumberger
}
