% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taunus_plot_porbe_plots.R
\name{taunus_plot_porbe_plots}
\alias{taunus_plot_porbe_plots}
\title{mapview plot of the locations of the taunus porbe data}
\usage{
taunus_plot_porbe_plots(path)
}
\arguments{
\item{path}{path to the local saved}
}
\value{
A mapview output
}
\description{
mapview plot of the locations of the taunus porbe data
}
\examples{
taunus_plot_porbe_plots()
\dontrun{
taunus_plot_porbe_plots()
}
}
\author{
Maiken Baumberger
}
