% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{test}
\alias{test}
\title{this is a test function}
\usage{
test(x)
}
\arguments{
\item{x}{whatever}
}
\value{
just a print
}
\description{
this is a test function
}
\examples{
\dontrun{
test("hallo")
}
}
\author{
Hanna Meyer
}
