% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test2.R
\name{test2}
\alias{test2}
\title{this is a second test function}
\usage{
test2(x)
}
\arguments{
\item{dateiname2}{whatever}
}
\value{
just a print
}
\description{
this is a description
}
\examples{
\dontrun{
test("hallo")
}
}
\author{
Hanna Meyer
}
