% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_probe_meta_data_longterm.R
\name{load_probe_meta_data_longterm}
\alias{load_probe_meta_data_longterm}
\title{load probe meta data longterm from a local folder into r}
\usage{
load_probe_meta_data_longterm(path)
}
\arguments{
\item{path}{path to the local saved}
}
\value{

}
\description{
load probe meta data longterm from a local folder into r
}
\examples{
\dontrun{
load_probe_meta_data_longterm
}
}
\author{
Maiken Baumberger
}
