% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPlots.R
\name{plotPlots}
\alias{plotPlots}
\title{Plot the Carbon4D plots}
\usage{
plotPlots(url)
}
\arguments{
\item{url}{URL of the Carbon-4D data base}
}
\value{
A mapview output
}
\description{
Plot the Carbon4D plots
}
\examples{
\dontrun{
plotPlots()
}
}
\author{
Hanna Meyer
}
